+++ src/main.c
@@ -6,27 +6,11 @@
 /*   By: vipereir <vipereir@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/02 14:03:30 by rgorki            #+#    #+#             */
-/*   Updated: 2023/05/06 13:53:40 by vipereir         ###   ########.fr       */
+/*   Updated: 2023/05/06 14:18:02 by vipereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "header/cub3d.h"
-
-int	get_key(int keycode, t_game *game)
-{
-	printf("key: %i\n", keycode);
-	if (keycode == KEY_ESC)
-		close_game(game);
-	if (keycode == KEY_UP)
-		move_forward(game);
-	if (keycode == KEY_DOWN)
-		move_backward(game);
-	if (keycode == CAM_ARROW_RIGHT)
-		rotate_camera_right(game);
-	if (keycode == CAM_ARROW_LEFT)
-		rotate_camera_left(game);
-	return (0);
-}
 
 int	main(int argc, char **argv)
 {
