+++ src/player_movement/wasd_mov.c
@@ -6,7 +6,7 @@
 /*   By: vipereir <vipereir@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/06 12:09:33 by vipereir          #+#    #+#             */
-/*   Updated: 2023/05/06 13:46:13 by vipereir         ###   ########.fr       */
+/*   Updated: 2023/05/06 14:12:14 by vipereir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,12 +30,20 @@ void    move_backward(t_game *game)
 	raycasting(game);
 }
 
-/* void    move_left(t_game *game)
+void    move_left(t_game *game)
 {
-
+	if (game->map.map_square[(int)(game->player.posY)][(int)(game->player.posX - game->player.planeX * game->player.moveSpeed)] == '0')
+		game->player.posX -= game->player.planeX * game->player.moveSpeed;
+	if (game->map.map_square[(int)(game->player.posY - game->player.planeY * game->player.moveSpeed)][(int)(game->player.posX)] == '0')
+		game->player.posY -= game->player.planeY * game->player.moveSpeed;
+	raycasting(game);
 }
 
 void    move_right(t_game *game)
 {
-
-} */
+	if (game->map.map_square[(int)(game->player.posY)][(int)(game->player.posX + game->player.planeX * game->player.moveSpeed)] == '0')
+		game->player.posX += game->player.planeX * game->player.moveSpeed;
+	if (game->map.map_square[(int)(game->player.posY + game->player.planeY * game->player.moveSpeed)][(int)(game->player.posX)] == '0')
+		game->player.posY += game->player.planeY * game->player.moveSpeed;
+	raycasting(game);
+}
